Add 'clone' and 'echo' to the list of known PHP statements to avoid the use of parenthesis. Move the sniff to a proper file and remove obsolete code.
